+++ packages/cli/src/commands/finalize.js
@@ -1,0 +1,1 @@
+export default async function finalizeCmd() { console.log('finalize'); }